backend/requirements.txt
# Consultation Management System - Backend Dependencies

# FastAPI and Web Server
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.12

# Pydantic for data validation
pydantic>=2.10.0
pydantic-settings>=2.7.0

# Supabase client (for production)
supabase==2.25.0

# Environment variables
python-dotenv>=1.0.0

# CORS and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Rate limiting (Redis storage makes limits shared across all workers/instances)
slowapi>=0.1.9
redis>=5.0.0

# Date and time utilities
python-dateutil>=2.8.2

# PDF generation for invoice attachments
reportlab>=4.0.0

# Testing (optional, for development)
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.28.0

# Code quality (optional, for development)
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
